phase 2

+++ public/popupScript.js
@@ -1,14 +1,14 @@
 const expirationMins = 10;
 
-function initPopup() {
+function initPopup(text) {
     const popupWrapper = document.createElement('div')
     popupWrapper.setAttribute('class', 'modal')
     popupWrapper.setAttribute('id', 'popup')
     popupWrapper.addEventListener('click', hidePopup)
 
     const content = document.createElement('div')
-    content.setAttribute('class','modal-content')
-    
+    content.setAttribute('class', 'modal-content')
+
     const confirmationButton = document.createElement('button')
     confirmationButton.addEventListener('click', handleConfirmation)
     confirmationButton.setAttribute('id', 'confirm')
@@ -16,6 +16,9 @@ function initPopup() {
 
     const message = document.createElement('div')
     message.innerHTML = "<h2>Please confirm that you read this.</h2>"
+    const message2 = document.createElement('div');
+    message2.setAttribute('id', 'message2')
+    message2.innerHTML = text;
 
     const closeButton = document.createElement('span')
     closeButton.setAttribute('id', 'close')
@@ -24,6 +27,7 @@ function initPopup() {
 
     content.appendChild(closeButton)
     content.appendChild(message)
+    content.appendChild(message2)
     content.appendChild(confirmationButton)
     popupWrapper.appendChild(content)
 
@@ -38,9 +42,20 @@ function hidePopup() {
     document.querySelector('#popup').style.display = "none"
 }
 
-function handleConfirmation() {
-    localStorage.setItem('popup_confirm_timestamp', Date.now().toString());
-    hidePopup()
+function handleConfirmation(e) {
+    // stop propagation to #popup which would trigger hidePopup()
+    e.stopPropagation();
+
+    fetch('/popup/confirmation', {
+            method: 'POST'
+        }
+    ).then(async res => {
+        const jsonResponse = await res.json();
+        if (jsonResponse.confirmationTracked) {
+            localStorage.setItem('popup_confirm_timestamp', Date.now().toString());
+            hidePopup();
+        }
+    }).catch(console.error);
 }
 
 function isPopupRequired() {
@@ -56,9 +71,20 @@ function isPopupRequired() {
     return lastConfirmationTimestamp + expiration < Date.now();
 }
 
-document.addEventListener('DOMContentLoaded', function(){
-    if (isPopupRequired()) {
-        initPopup()
-        showPopup()
+function displayPopup() {
+    if (!isPopupRequired()) {
+        return;
     }
+
+    fetch('/popup').then(async res => {
+        const jsonResponse = await res.json();
+        if (jsonResponse.message) {
+            initPopup(jsonResponse.message)
+            showPopup()
+        }
+    }).catch(console.error);
+}
+
+document.addEventListener('DOMContentLoaded', function () {
+    displayPopup();
 }, false);
